+++ src/components/layouts/Wide+Side.ts
@@ -1,21 +1,11 @@
-import './common.scss';
 import './Wide.scss';
 import './Narrow.scss';
 
-import '../Head';
-import {Block} from '../../modules/Block';
+import {Layout} from './Layout';
 
-export class WideLayoutWithSidebar extends Block {
+export class WideLayoutWithSidebar extends Layout {
   constructor(props: Record<string, unknown>) {
     super(props);
-    this.on('mounted', () => {
-      const popup = document.querySelector('.popup') as HTMLElement;
-      popup.addEventListener('click', (event: Event) => {
-        if (event.target === popup) {
-          popup.style.display = 'none';
-        }
-      });
-    });
   }
   render(): string {
     return `
