+++ src/components/layouts/common.scss
@@ -109,6 +109,11 @@ body {
   }
 }
 
+a {
+  color: inherit;
+  text-decoration: inherit;
+}
+
 ul,
 li {
   margin: 0;
