+++ src/pages/404.ts
@@ -1,10 +1,12 @@
 import {WideLayout} from '../components/layouts/Wide';
 import {ErrorStatusMsg} from '../components/ErrorStatusMsg';
+import {Link} from '../components/Link';
 
 const view = new WideLayout({
   title: '404',
   comment: 'Не туда попали',
   contents: ErrorStatusMsg,
+  navLink: new Link({url: 'chats.html', label: 'Назад к чатам'}),
 });
 
 export {view};
