+++ src/pages/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <script type="module">
   import {Router} from '../modules/Router';
-  import {view as IndexPage} from './index.ts';
+  import {view as IndexPage} from './main';
   import {view as AuthPage} from './auth';
   import {view as RegisterPage} from './register';
   import {view as ProfilePage} from './profile';
